pyproject.toml
[project]
name = "ppml"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
# requires-python = ">=3.10,<4"
[tool.poetry.dependencies]
python = ">=3.10,<4.0"
fastapi = ">=0.135.1,<0.136.0"
pydantic = ">=2.12.5,<3.0.0"
python-dotenv = ">=1.2.2,<2.0.0"
psutil = ">=7.2.2,<8.0.0"
uvicorn = ">=0.42.0,<0.43.0"
pandas = "<3"
motor = "^3.7.1"
pymongo = "^4.16.0"
requests = "^2.33.0"
tortoise-orm = {extras = ["asyncmy"], version = "^1.1.7"}
commonx = {version = "0.0.1a3", allow-prereleases = true, source = "test"}
xolo = {version = "0.0.11a2", allow-prereleases = true, source = "test"}
httpx = "^0.28.1"
aiomysql = "^0.3.2"
roryclient = {version = "0.1.13", allow-prereleases = true, source = "test"}


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
coverage = "^7.13.5"
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
zensical = "^0.0.29"

[[tool.poetry.source]]
name = "test"
url = "https://test.pypi.org/simple/"
priority = "explicit"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_test_loop_scope = "session"